Changed type of input at method compute() @ Fibonacci.java to let user compute any value of the fibonacci sequence.

+++ src/main/java/org/jlbp/fibonacci/Fibonacci.java
@@ -10,7 +10,7 @@ import java.math.BigInteger;
  */
 
 public class Fibonacci {
-    public BigInteger compute(int n) {
+    public BigInteger compute(long n) {
         if(n < 0){
             throw new RuntimeException("ERROR: Input cannot be negative.");
         }
